Move type to types folder
